Arreglos al controlador de Cliente, a su modelo y al modelo Persona

+++ controller/clientController.js
@@ -1,5 +1,6 @@
 const Client = require('../model/Client')
 const Person = require('../model/Person')
+const Barber = require('../model/Barber')
 const { validationResult } = require('express-validator')
 
 exports.createClient = async (req, res) => {
@@ -12,19 +13,26 @@ exports.createClient = async (req, res) => {
     try {
         const { name, lastName, dni, phone } = req.body
 
-        const existsClient = await Client.findOne({ where: {dni: dni} })
-        if (existsClient) {
-            return res.status(400).send(`There is a client with the DNI: ${dni}`)
-        }
+        console.error(`${name} ${lastName} ${dni} ${phone}`)
 
         // Se verifica si ya existe la persona.
         // Si un barbero quiere cortarse el pelo, entonces se verifica si ya existen sus datos personales
-        const existsPerson = await Person.findByPk(dni)
-        if (!existsPerson) {
-            await Person.create({name, lastName, dni, phone})
+        const existsPerson = await Person.findOne({ where: { dni: dni} })
+        if (existsPerson) {
+            
+            const existClient = await Client.findOne({ where: { person_id: existsPerson.id } })
+            if (existClient) {
+                res.status(400).send(`There is a client with DNI: ${existsPerson.dni}`)
+            }
         }
 
-        await Client.create({dni})
+        const person = await Person.create({
+            name: name,
+            lastName: lastName,
+            dni: dni,
+            phone: phone
+        })
+        await Client.create({person_id: person.id})
 
         res.status(200).send(`New client added: ${lastName}, ${name}`)
 
@@ -38,7 +46,7 @@ exports.getClients = async (req, res) => {
     try {
         const clients = await Client.findAll({
             attributes: { exclude: ['createdAt', 'updatedAt' ] },
-            include: { model: Person, attributes: { exclude: ['createdAt', 'updatedAt', 'dni'] } }
+            include: { model: Person, attributes: { exclude: ['createdAt', 'updatedAt'] } }
         })
 
         res.json({clients})
@@ -55,7 +63,7 @@ exports.getClient = async (req, res) => {
 
         const client = await Client.findByPk(idClient, {
             attributes: { exclude: ['createdAt', 'updatedAt' ] },
-            include: { model: Person, attributes: { exclude: ['createdAt', 'updatedAt', 'dni'] }}
+            include: { model: Person, attributes: { exclude: ['createdAt', 'updatedAt'] }}
         })
         
         if (!client) {
@@ -77,23 +85,26 @@ exports.updateClient = async (req, res) => {
     }
 
     try {
+        // Se extraen los datos del body de la request
         const { name, lastName, dni, phone } = req.body
 
-        // Obtener id del cliente de los parámetros
-        const idClient = req.params.id
+        // Obtener id del cliente a modificar
+        const id = req.params.id
 
-        const client = await Client.findByPk(idClient)
+        // Validar que exista el cliente a modificar
+        const client = await Client.findByPk(id)
         if (!client) {
-            return res.status(400).send(`Client with id ${idClient} not found`)
+            return res.status(400).send(`Client not found`)
         }
 
-        // Se verifica si existe una persona con el dni nuevo
-        const person = await Person.findByPk(dni)
-        if (person) {
-            return res.status(400).send(`Person with DNI ${dni} not found`)
+        // Se verifica si existe una persona con el nuevo DNI
+        const person = await Person.findByPk(client.person_id)
+        if (person && !person.dni === dni) {
+            return res.status(400).send(`There is a person with the selected dni`)
         }
 
-        const clientPerson = await Person.findByPk(client.dni)
+        // Actualizar los datos personales del cliente
+        const clientPerson = await Person.findByPk(client.person_id)
         clientPerson.set({
             name: name,
             lastName: lastName,
@@ -102,12 +113,7 @@ exports.updateClient = async (req, res) => {
         })
         await clientPerson.save()
 
-        client.set({
-            dni: dni
-        })
-        await client.save()
-
-        res.status(200).send(`${clientPerson.lastName}, ${clientPerson.name} updated`)
+        res.status(200).send(`Client updated`)
     } catch (error) {
         console.log(error)
         return res.status(500).send('There was a server error')
@@ -122,10 +128,14 @@ exports.deleteClient = async (req, res) => {
         if (!client) {
             return res.status(404).send(`The client with id ${id} was not found`)
         }
-        
-        const person = Person.findByPk(client.dni)
-        await person.destroy()
+
+        const person = await Person.findByPk(client.person_id)
+        if (!person) {
+            return res.status(404).send(`The person with id ${id} was not found`)
+        }
+
         await client.destroy()
+        await person.destroy()
 
         res.status(200).send('Client deleted')
